add retry api Inner Class (SqlQueryImpl/SqlUpdateImpl/ProcedureImpl) to public Class(default scope)
