+++ src/main/java/jp/co/future/uroborosql/AbstractAgent.java
@@ -1,35 +1,25 @@
 package jp.co.future.uroborosql;
 
-import java.io.InputStream;
-import java.io.Reader;
 import java.sql.CallableStatement;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.SQLType;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.Stream;
 
 import jp.co.future.uroborosql.config.SqlConfig;
 import jp.co.future.uroborosql.connection.ConnectionSupplier;
 import jp.co.future.uroborosql.context.SqlContext;
-import jp.co.future.uroborosql.converter.MapResultSetConverter;
-import jp.co.future.uroborosql.converter.ResultSetConverter;
 import jp.co.future.uroborosql.coverage.CoverageData;
 import jp.co.future.uroborosql.coverage.CoverageHandler;
-import jp.co.future.uroborosql.exception.DataNotFoundException;
 import jp.co.future.uroborosql.exception.UroborosqlRuntimeException;
 import jp.co.future.uroborosql.filter.SqlFilterManager;
 import jp.co.future.uroborosql.fluent.Procedure;
 import jp.co.future.uroborosql.fluent.SqlQuery;
 import jp.co.future.uroborosql.fluent.SqlUpdate;
-import jp.co.future.uroborosql.parameter.Parameter;
 import jp.co.future.uroborosql.parser.ContextTransformer;
 import jp.co.future.uroborosql.parser.SqlParser;
 import jp.co.future.uroborosql.parser.SqlParserImpl;
@@ -38,7 +28,6 @@ import jp.co.future.uroborosql.tx.LocalTransactionManager;
 import jp.co.future.uroborosql.tx.SQLRunnable;
 import jp.co.future.uroborosql.tx.SQLSupplier;
 import jp.co.future.uroborosql.tx.TransactionManager;
-import jp.co.future.uroborosql.utils.CaseFormat;
 import jp.co.future.uroborosql.utils.StringFunction;
 
 import org.apache.commons.lang3.StringUtils;
@@ -116,7 +105,8 @@ public abstract class AbstractAgent implements SqlAgent {
 			try {
 				handler = (CoverageHandler) Class.forName(sqlCoverageClassName).newInstance();
 			} catch (Exception ex) {
-				LOG.warn("Failed to generate CoverageHandler class. Class：{}, Cause：{}", sqlCoverageClassName, ex.getMessage());
+				LOG.warn("Failed to generate CoverageHandler class. Class：{}, Cause：{}", sqlCoverageClassName,
+						ex.getMessage());
 			}
 		}
 
@@ -677,729 +667,6 @@ public abstract class AbstractAgent implements SqlAgent {
 		// プロパティ設定
 		applyProperties(stmt);
 		return stmt;
-	}
-
-	/**
-	 * SqlQuery実装
-	 *
-	 * @author H.Sugimoto
-	 */
-	private final class SqlQueryImpl implements SqlQuery {
-		private final SqlAgent agent;
-		private final SqlContext context;
-
-		/**
-		 * コンストラクタ
-		 *
-		 * @param agent SqlAgent
-		 * @param context SqlContext
-		 */
-		private SqlQueryImpl(final SqlAgent agent, final SqlContext context) {
-			this.agent = agent;
-			this.context = context;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#paramList(java.lang.String, java.lang.Object[])
-		 */
-		@Override
-		public SqlQuery paramList(final String paramName, final Object... value) {
-			context.paramList(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#paramMap(Map<String, Object>)
-		 */
-		@Override
-		public SqlQuery paramMap(final Map<String, Object> paramMap) {
-			if (paramMap != null) {
-				paramMap.forEach((k, v) -> param(k, v));
-			}
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(java.lang.String, java.lang.Object, int)
-		 */
-		@Override
-		public SqlQuery param(final String paramName, final Object value, final int sqlType) {
-			context.param(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(java.lang.String, java.lang.Object, java.sql.SQLType)
-		 */
-		@Override
-		public SqlQuery param(final String paramName, final Object value, final SQLType sqlType) {
-			context.param(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(java.lang.String, java.lang.Object)
-		 */
-		@Override
-		public SqlQuery param(final String paramName, final Object value) {
-			context.param(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(jp.co.future.uroborosql.parameter.Parameter)
-		 */
-		@Override
-		public SqlQuery param(final Parameter param) {
-			context.param(param);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#outParam(java.lang.String, int)
-		 */
-		@Override
-		public SqlQuery outParam(final String paramName, final int sqlType) {
-			context.outParam(paramName, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#outParam(java.lang.String, java.sql.SQLType)
-		 */
-		@Override
-		public SqlQuery outParam(final String paramName, final SQLType sqlType) {
-			context.outParam(paramName, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#inOutParam(java.lang.String, java.lang.Object, int)
-		 */
-		@Override
-		public SqlQuery inOutParam(final String paramName, final Object value, final int sqlType) {
-			context.inOutParam(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#inOutParam(java.lang.String, java.lang.Object,
-		 *      java.sql.SQLType)
-		 */
-		@Override
-		public SqlQuery inOutParam(final String paramName, final Object value, final SQLType sqlType) {
-			context.inOutParam(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#characterStreamParam(java.lang.String, java.io.Reader, int)
-		 */
-		@Override
-		public SqlQuery characterStreamParam(final String paramName, final Reader value, final int len) {
-			context.characterStreamParam(paramName, value, len);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#characterStreamParam(java.lang.String, java.io.Reader)
-		 */
-		@Override
-		public SqlQuery characterStreamParam(final String paramName, final Reader value) {
-			context.characterStreamParam(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#binaryStreamParam(java.lang.String, java.io.InputStream, int)
-		 */
-		@Override
-		public SqlQuery binaryStreamParam(final String paramName, final InputStream value, final int len) {
-			context.binaryStreamParam(paramName, value, len);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#binaryStreamParam(java.lang.String, java.io.InputStream)
-		 */
-		@Override
-		public SqlQuery binaryStreamParam(final String paramName, final InputStream value) {
-			context.binaryStreamParam(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#asciiStreamParam(java.lang.String, java.io.InputStream, int)
-		 */
-		@Override
-		public SqlQuery asciiStreamParam(final String paramName, final InputStream value, final int len) {
-			context.asciiStreamParam(paramName, value, len);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#asciiStreamParam(java.lang.String, java.io.InputStream)
-		 */
-		@Override
-		public SqlQuery asciiStreamParam(final String paramName, final InputStream value) {
-			context.asciiStreamParam(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlQuery#stream()
-		 */
-		@Override
-		public Stream<Map<String, Object>> stream() throws SQLException {
-			return stream(new MapResultSetConverter());
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlQuery#stream(jp.co.future.uroborosql.converter.ResultSetConverter)
-		 */
-		@Override
-		public <T> Stream<T> stream(final ResultSetConverter<T> converter) throws SQLException {
-			return agent.query(context, converter);
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlQuery#resultSet()
-		 */
-		@Override
-		public ResultSet resultSet() throws SQLException {
-			return agent.query(context);
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlQuery#first()
-		 */
-		@Override
-		public Map<String, Object> first() throws DataNotFoundException, SQLException {
-			return first(CaseFormat.SnakeCase);
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlQuery#first(jp.co.future.uroborosql.utils.CaseFormat)
-		 */
-		@Override
-		public Map<String, Object> first(final CaseFormat caseFormat) throws DataNotFoundException, SQLException {
-			Optional<Map<String, Object>> first = stream(new MapResultSetConverter(caseFormat)).findFirst();
-			return first.orElseThrow(() -> new DataNotFoundException("query result is empty."));
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlQuery#collect(jp.co.future.uroborosql.utils.CaseFormat)
-		 */
-		@Override
-		public List<Map<String, Object>> collect(final CaseFormat caseFormat) throws SQLException {
-			return agent.query(context, caseFormat);
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlQuery#collect()
-		 */
-		@Override
-		public List<Map<String, Object>> collect() throws SQLException {
-			return collect(CaseFormat.SnakeCase);
-		}
-
-	}
-
-	/**
-	 * SqlUpdate実装
-	 *
-	 * @author H.Sugimoto
-	 */
-	private final class SqlUpdateImpl implements SqlUpdate {
-		private final SqlAgent agent;
-		private final SqlContext context;
-		/** バッチ処理を行うかどうか */
-		boolean batch = false;
-
-		/**
-		 * コンストラクタ
-		 *
-		 * @param agent SqlAgent
-		 * @param context SqlContext
-		 */
-		private SqlUpdateImpl(final SqlAgent agent, final SqlContext context) {
-			this.agent = agent;
-			this.context = context;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#paramList(java.lang.String, java.lang.Object[])
-		 */
-		@Override
-		public SqlUpdate paramList(final String paramName, final Object... value) {
-			context.paramList(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#paramMap(Map<String, Object>)
-		 */
-		@Override
-		public SqlUpdate paramMap(final Map<String, Object> paramMap) {
-			if (paramMap != null) {
-				paramMap.forEach((k, v) -> param(k, v));
-			}
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(java.lang.String, java.lang.Object, int)
-		 */
-		@Override
-		public SqlUpdate param(final String paramName, final Object value, final int sqlType) {
-			context.param(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(java.lang.String, java.lang.Object, java.sql.SQLType)
-		 */
-		@Override
-		public SqlUpdate param(final String paramName, final Object value, final SQLType sqlType) {
-			context.param(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(java.lang.String, java.lang.Object)
-		 */
-		@Override
-		public SqlUpdate param(final String paramName, final Object value) {
-			context.param(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(jp.co.future.uroborosql.parameter.Parameter)
-		 */
-		@Override
-		public SqlUpdate param(final Parameter param) {
-			context.param(param);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#outParam(java.lang.String, int)
-		 */
-		@Override
-		public SqlUpdate outParam(final String paramName, final int sqlType) {
-			context.outParam(paramName, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#outParam(java.lang.String, java.sql.SQLType)
-		 */
-		@Override
-		public SqlUpdate outParam(final String paramName, final SQLType sqlType) {
-			context.outParam(paramName, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#inOutParam(java.lang.String, java.lang.Object, int)
-		 */
-		@Override
-		public SqlUpdate inOutParam(final String paramName, final Object value, final int sqlType) {
-			context.inOutParam(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#inOutParam(java.lang.String, java.lang.Object,
-		 *      java.sql.SQLType)
-		 */
-		@Override
-		public SqlUpdate inOutParam(final String paramName, final Object value, final SQLType sqlType) {
-			context.inOutParam(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#characterStreamParam(java.lang.String, java.io.Reader, int)
-		 */
-		@Override
-		public SqlUpdate characterStreamParam(final String paramName, final Reader value, final int len) {
-			context.characterStreamParam(paramName, value, len);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#characterStreamParam(java.lang.String, java.io.Reader)
-		 */
-		@Override
-		public SqlUpdate characterStreamParam(final String paramName, final Reader value) {
-			context.characterStreamParam(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#binaryStreamParam(java.lang.String, java.io.InputStream, int)
-		 */
-		@Override
-		public SqlUpdate binaryStreamParam(final String paramName, final InputStream value, final int len) {
-			context.binaryStreamParam(paramName, value, len);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#binaryStreamParam(java.lang.String, java.io.InputStream)
-		 */
-		@Override
-		public SqlUpdate binaryStreamParam(final String paramName, final InputStream value) {
-			context.binaryStreamParam(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#asciiStreamParam(java.lang.String, java.io.InputStream, int)
-		 */
-		@Override
-		public SqlUpdate asciiStreamParam(final String paramName, final InputStream value, final int len) {
-			context.asciiStreamParam(paramName, value, len);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#asciiStreamParam(java.lang.String, java.io.InputStream)
-		 */
-		@Override
-		public SqlUpdate asciiStreamParam(final String paramName, final InputStream value) {
-			context.asciiStreamParam(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlUpdate#addBatch()
-		 */
-		@Override
-		public SqlUpdate addBatch() {
-			context.addBatch();
-			batch = true;
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlUpdate#count()
-		 */
-		@Override
-		public int count() throws SQLException {
-			if (batch) {
-				throw new IllegalStateException("すでにaddBatch()でパラメータが設定されているため、batch()を呼び出してください");
-			}
-			return agent.update(context);
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlUpdate#batch()
-		 */
-		@Override
-		public int[] batch() throws SQLException {
-			if (!batch) {
-				addBatch();
-			}
-			return agent.batch(context);
-		}
-
-	}
-
-	/**
-	 * Procedure実装
-	 *
-	 * @author H.Sugimoto
-	 */
-	private final class ProcedureImpl implements Procedure {
-		private final SqlAgent agent;
-		private final SqlContext context;
-
-		/**
-		 * コンストラクタ
-		 *
-		 * @param agent SqlAgent
-		 * @param context SqlContext
-		 */
-		private ProcedureImpl(final SqlAgent agent, final SqlContext context) {
-			this.agent = agent;
-			this.context = context;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#paramList(java.lang.String, java.lang.Object[])
-		 */
-		@Override
-		public Procedure paramList(final String paramName, final Object... value) {
-			context.paramList(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#paramMap(Map<String, Object>)
-		 */
-		@Override
-		public Procedure paramMap(final Map<String, Object> paramMap) {
-			if (paramMap != null) {
-				paramMap.forEach((k, v) -> param(k, v));
-			}
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(java.lang.String, java.lang.Object, int)
-		 */
-		@Override
-		public Procedure param(final String paramName, final Object value, final int sqlType) {
-			context.param(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(java.lang.String, java.lang.Object, java.sql.SQLType)
-		 */
-		@Override
-		public Procedure param(final String paramName, final Object value, final SQLType sqlType) {
-			context.param(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(java.lang.String, java.lang.Object)
-		 */
-		@Override
-		public Procedure param(final String paramName, final Object value) {
-			context.param(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#param(jp.co.future.uroborosql.parameter.Parameter)
-		 */
-		@Override
-		public Procedure param(final Parameter param) {
-			context.param(param);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#outParam(java.lang.String, int)
-		 */
-		@Override
-		public Procedure outParam(final String paramName, final int sqlType) {
-			context.outParam(paramName, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#outParam(java.lang.String, java.sql.SQLType)
-		 */
-		@Override
-		public Procedure outParam(final String paramName, final SQLType sqlType) {
-			context.outParam(paramName, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#inOutParam(java.lang.String, java.lang.Object, int)
-		 */
-		@Override
-		public Procedure inOutParam(final String paramName, final Object value, final int sqlType) {
-			context.inOutParam(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#inOutParam(java.lang.String, java.lang.Object,
-		 *      java.sql.SQLType)
-		 */
-		@Override
-		public Procedure inOutParam(final String paramName, final Object value, final SQLType sqlType) {
-			context.inOutParam(paramName, value, sqlType);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#characterStreamParam(java.lang.String, java.io.Reader, int)
-		 */
-		@Override
-		public Procedure characterStreamParam(final String paramName, final Reader value, final int len) {
-			context.characterStreamParam(paramName, value, len);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#characterStreamParam(java.lang.String, java.io.Reader)
-		 */
-		@Override
-		public Procedure characterStreamParam(final String paramName, final Reader value) {
-			context.characterStreamParam(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#binaryStreamParam(java.lang.String, java.io.InputStream, int)
-		 */
-		@Override
-		public Procedure binaryStreamParam(final String paramName, final InputStream value, final int len) {
-			context.binaryStreamParam(paramName, value, len);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#binaryStreamParam(java.lang.String, java.io.InputStream)
-		 */
-		@Override
-		public Procedure binaryStreamParam(final String paramName, final InputStream value) {
-			context.binaryStreamParam(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#asciiStreamParam(java.lang.String, java.io.InputStream, int)
-		 */
-		@Override
-		public Procedure asciiStreamParam(final String paramName, final InputStream value, final int len) {
-			context.asciiStreamParam(paramName, value, len);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.SqlFluent#asciiStreamParam(java.lang.String, java.io.InputStream)
-		 */
-		@Override
-		public Procedure asciiStreamParam(final String paramName, final InputStream value) {
-			context.asciiStreamParam(paramName, value);
-			return this;
-		}
-
-		/**
-		 * {@inheritDoc}
-		 *
-		 * @see jp.co.future.uroborosql.fluent.Procedure#call()
-		 */
-		@Override
-		public Map<String, Object> call() throws SQLException {
-			return agent.procedure(context);
-		}
 	}
 
 }
