+++ src/main/java/jp/co/future/uroborosql/SqlAgentImpl.java
@@ -101,14 +101,14 @@ public class SqlAgentImpl extends AbstractAgent {
 			do {
 				try {
 					ResultSet rs = getSqlFilterManager().doQuery(sqlContext, stmt, stmt.executeQuery());
-					sqlContext.contextAttrs().put("__retryCount", loopCount);
 					return rs;
 				} catch (SQLException ex) {
 					if (maxRetryCount > loopCount) {
 						String errorCode = Integer.toString(ex.getErrorCode());
 						if (getSqlRetryCodes().contains(errorCode)) {
 							if (LOG.isDebugEnabled()) {
-								LOG.debug(String.format("Caught the error code to be retried.(%d times). Retry after %,3d ms.",
+								LOG.debug(String.format(
+										"Caught the error code to be retried.(%d times). Retry after %,3d ms.",
 										loopCount + 1, retryWaitTime));
 							}
 							if (retryWaitTime > 0) {
@@ -124,6 +124,8 @@ public class SqlAgentImpl extends AbstractAgent {
 					} else {
 						throw ex;
 					}
+				} finally {
+					sqlContext.contextAttrs().put("__retryCount", loopCount);
 				}
 			} while (maxRetryCount > loopCount++);
 			return null;
@@ -251,14 +253,14 @@ public class SqlAgentImpl extends AbstractAgent {
 			do {
 				try {
 					int result = getSqlFilterManager().doUpdate(sqlContext, stmt, stmt.executeUpdate());
-					sqlContext.contextAttrs().put("__retryCount", loopCount);
 					return result;
 				} catch (SQLException ex) {
 					if (maxRetryCount > loopCount) {
 						String errorCode = Integer.toString(ex.getErrorCode());
 						if (getSqlRetryCodes().contains(errorCode)) {
 							if (LOG.isDebugEnabled()) {
-                                LOG.debug(String.format("Caught the error code to be retried.(%d times). Retry after %,3d ms.",
+								LOG.debug(String.format(
+										"Caught the error code to be retried.(%d times). Retry after %,3d ms.",
 										loopCount + 1, retryWaitTime));
 							}
 							if (retryWaitTime > 0) {
@@ -274,6 +276,8 @@ public class SqlAgentImpl extends AbstractAgent {
 					} else {
 						throw ex;
 					}
+				} finally {
+					sqlContext.contextAttrs().put("__retryCount", loopCount);
 				}
 			} while (maxRetryCount > loopCount++);
 			return 0;
@@ -348,14 +352,14 @@ public class SqlAgentImpl extends AbstractAgent {
 			do {
 				try {
 					int[] result = getSqlFilterManager().doBatch(sqlContext, stmt, stmt.executeBatch());
-					sqlContext.contextAttrs().put("__retryCount", loopCount);
 					return result;
 				} catch (SQLException ex) {
 					if (maxRetryCount > loopCount) {
 						String errorCode = Integer.toString(ex.getErrorCode());
 						if (getSqlRetryCodes().contains(errorCode)) {
 							if (LOG.isDebugEnabled()) {
-								LOG.debug(String.format("Caught the error code to be retried.(%d times). Retry after %,3d ms.",
+								LOG.debug(String.format(
+										"Caught the error code to be retried.(%d times). Retry after %,3d ms.",
 										loopCount + 1, retryWaitTime));
 							}
 							if (retryWaitTime > 0) {
@@ -371,6 +375,8 @@ public class SqlAgentImpl extends AbstractAgent {
 					} else {
 						throw ex;
 					}
+				} finally {
+					sqlContext.contextAttrs().put("__retryCount", loopCount);
 				}
 			} while (maxRetryCount > loopCount++);
 			return null;
@@ -454,7 +460,8 @@ public class SqlAgentImpl extends AbstractAgent {
 						String errorCode = Integer.toString(ex.getErrorCode());
 						if (getSqlRetryCodes().contains(errorCode)) {
 							if (LOG.isDebugEnabled()) {
-								LOG.debug(String.format("Caught the error code to be retried.(%d times). Retry after %,3d ms.",
+								LOG.debug(String.format(
+										"Caught the error code to be retried.(%d times). Retry after %,3d ms.",
 										loopCount + 1, retryWaitTime));
 							}
 							if (retryWaitTime > 0) {
@@ -470,6 +477,8 @@ public class SqlAgentImpl extends AbstractAgent {
 					} else {
 						throw ex;
 					}
+				} finally {
+					sqlContext.contextAttrs().put("__retryCount", loopCount);
 				}
 			} while (maxRetryCount > loopCount++);
 		} catch (SQLException ex) {
@@ -520,8 +529,10 @@ public class SqlAgentImpl extends AbstractAgent {
 
 		if (LOG.isErrorEnabled() && isOutputExceptionLog()) {
 			StringBuilder builder = new StringBuilder();
-			builder.append(System.lineSeparator()).append("Exception occurred in SQL execution.").append(System.lineSeparator());
-			builder.append("Executed SQL[").append(sqlContext.getExecutableSql()).append("]").append(System.lineSeparator());
+			builder.append(System.lineSeparator()).append("Exception occurred in SQL execution.")
+					.append(System.lineSeparator());
+			builder.append("Executed SQL[").append(sqlContext.getExecutableSql()).append("]")
+					.append(System.lineSeparator());
 			if (sqlContext instanceof SqlContextImpl) {
 				Parameter[] bindParameters = ((SqlContextImpl) sqlContext).getBindParameters();
 				for (int i = 0; i < bindParameters.length; i++) {
